internal/factory/container: make (*container).SelinuxLabel platform-specific
This allows for a no-op stub on FreeBSD which does not have SELinux.
Note: FreeBSD does have mandatory access control labels but I am not
intending to try to implement them for containers just yet.
