+++ internal/factory/container/container_freebsd.go
@@ -1,0 +1,5 @@
+package container
+
+func (c *container) SelinuxLabel(sboxLabel string) ([]string, error) {
+	return []string{}, nil
+}
